requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
numpy
scikit-learn
cryptography
python-dotenv
onnxruntime
scipy
